@@ -1,12 +1,18 @@
 plugins {
     id 'java'
-    id "com.diffplug.spotless" version "5.16.0"
-    id "org.sonarqube" version "3.3"
-    id "name.remal.sonarlint" version "1.5.0"
+    id "com.diffplug.spotless" version "6.25.0"
+    id "org.sonarqube" version "5.1.0.4882"
+    id "name.remal.sonarlint" version "4.2.2"
 }
 
 group 'org.togetherjava'
 version '1.0-SNAPSHOT'
+
+ext {
+    jooqVersion = '3.19.1'
+    jacksonVersion = '2.17.0'
+    chatGPTVersion = '0.18.0'
+}
 
 // Skips sonarlint during the build, useful for testing purposes.
 Boolean skipSonarlint = false
@@ -20,6 +26,14 @@ sonarqube {
         property "sonar.host.url", "https://sonarcloud.io"
     }
 }
+
+// Install git hooks
+task installLocalGitHook(type: Copy) {
+    from new File(rootProject.rootDir, 'scripts/pre-commit')
+    into new File(rootProject.rootDir, '.git/hooks')
+    fileMode 0775
+}
+build.dependsOn installLocalGitHook
 
 subprojects {
     apply plugin: "java"
@@ -37,35 +51,48 @@ subprojects {
     if (!skipSonarlint) {
         apply plugin: "name.remal.sonarlint"
     }
-
-    // sonarlint configuration, not to be confused with sonarqube/sonarcloud.
-    sonarlint {
-        excludes {
-            // Disables "Track uses of "TODO" tags" rule.
-            message 'java:S1135'
-
-            // Disables "Regular expressions should not overflow the stack" rule.
-            // https://sonarcloud.io/project/issues?id=Together-Java_TJ-Bot&issues=AXvleoajNC7zjjhmSufQ&open=AXvleoajNC7zjjhmSufQ
-            // https://sonarcloud.io/project/issues?id=Together-Java_TJ-Bot&issues=AXvleoX5NC7zjjhmSufP&open=AXvleoX5NC7zjjhmSufP
-            message 'java:S5998'
+    
+    java {
+        toolchain {
+            // Nails the Java-Version of every Subproject
+            languageVersion = JavaLanguageVersion.of(21)
         }
     }
 
-
-    compileJava {
-        // Makes spotlessApply task run on every compile/build.
-        dependsOn 'spotlessApply'
-
-        // Nails the Java-Version of every Subproject
-        sourceCompatibility = JavaVersion.VERSION_17
-        targetCompatibility = JavaVersion.VERSION_17
+    // sonarlint configuration, not to be confused with sonarqube/sonarcloud.
+    sonarLint {
+        rules {
+            disable(
+                    'java:S1135' // Disables "Track uses of "TO-DO" tags" rule.
+            )
+        }
     }
 
     spotless {
         java {
             // Excludes build folder since it contains generated java classes.
             targetExclude("build/**")
-            eclipse('4.19.0').configFile("${rootProject.rootDir}/meta/formatting/google-style-eclipse.xml")
+            endWithNewline()
+            removeUnusedImports()
+            // empty string '' covers all imports that aren't explicitly specified,
+            // we use it as catch-all for external dependencies like JDA
+            // '\\#` is prefix for static imports
+            importOrder('','org.togetherjava', 'javax', 'java', '\\#')
+            // TODO: pinning version because of spotless error https://github.com/diffplug/spotless/issues/1992
+            eclipse("4.31").configFile("${rootProject.rootDir}/meta/formatting/google-style-eclipse.xml")
         }
     }
+
+    test {
+        useJUnitPlatform()
+    }
+
+    compileJava {
+        options.encoding = "UTF-8"
+    }
+
+    compileTestJava {
+        options.encoding = "UTF-8"
+    }
+    
 }
