@@ -3,10 +3,12 @@ package org.togetherjava.tjbot.db;
 import org.flywaydb.core.Flyway;
 import org.jooq.DSLContext;
 import org.jooq.SQLDialect;
+import org.jooq.Table;
 import org.jooq.exception.DataAccessException;
 import org.jooq.impl.DSL;
 import org.sqlite.SQLiteConfig;
 import org.sqlite.SQLiteDataSource;
+
 import org.togetherjava.tjbot.db.util.CheckedConsumer;
 import org.togetherjava.tjbot.db.util.CheckedFunction;
 
@@ -23,6 +25,11 @@ import java.util.concurrent.locks.ReentrantLock;
  * Instances of this class are thread-safe and can be used to concurrently write to the database.
  */
 public final class Database {
+
+    static {
+        System.setProperty("org.jooq.no-logo", "true");
+        System.setProperty("org.jooq.no-tips", "true");
+    }
 
     private final DSLContext dslContext;
     /**
@@ -54,6 +61,22 @@ public final class Database {
     }
 
     /**
+     * Creates a new empty database that is hold in memory.
+     *
+     * @param tables the tables the database will hold if desired, otherwise null
+     * @return the created database
+     */
+    public static Database createMemoryDatabase(Table<?>... tables) {
+        try {
+            Database database = new Database("jdbc:sqlite:");
+            database.write(context -> context.ddl(tables).executeBatch());
+            return database;
+        } catch (SQLException e) {
+            throw new DatabaseException(e);
+        }
+    }
+
+    /**
      * Acquires read-only access to the database.
      *
      * @param action the action to apply to the DSL context, e.g. a query
@@ -71,12 +94,13 @@ public final class Database {
     }
 
     /**
-     * Acquires read-only access to the database.
+     * Acquires read-only access to the database and consumes the result directly.
      *
      * @param action the action that consumes the DSL context, e.g. a query
      * @throws DatabaseException if an error occurs in the given action
      */
-    public void read(CheckedConsumer<? super DSLContext, ? extends DataAccessException> action) {
+    public void readAndConsume(
+            CheckedConsumer<? super DSLContext, ? extends DataAccessException> action) {
         read(context -> {
             action.accept(context);
             // noinspection ReturnOfNull
@@ -85,14 +109,14 @@ public final class Database {
     }
 
     /**
-     * Acquires read and write access to the database.
+     * Acquires read and write access to the database and provides the computed result.
      *
      * @param action the action to apply to the DSL context, e.g. a query
      * @param <T> the type returned by the given action
      * @return the result returned by the given action
      * @throws DatabaseException if an error occurs in the given action
      */
-    public <T> T write(
+    public <T> T writeAndProvide(
             CheckedFunction<? super DSLContext, T, ? extends DataAccessException> action) {
         writeLock.lock();
         try {
@@ -111,7 +135,7 @@ public final class Database {
      * @throws DatabaseException if an error occurs in the given action
      */
     public void write(CheckedConsumer<? super DSLContext, ? extends DataAccessException> action) {
-        write(context -> {
+        writeAndProvide(context -> {
             action.accept(context);
             // noinspection ReturnOfNull
             return null;
@@ -141,13 +165,13 @@ public final class Database {
     }
 
     /**
-     * Acquires a transaction that can only read from the database.
+     * Acquires a transaction that can only read from the database and consumes the result directly.
      *
      * @param handler the handler that is executed within the context of the transaction. It has no
      *        return value.
      * @throws DatabaseException if an error occurs in the given handler function
      */
-    public void readTransaction(
+    public void readTransactionAndConsume(
             CheckedConsumer<? super DSLContext, ? extends DataAccessException> handler) {
         readTransaction(dsl -> {
             handler.accept(dsl);
@@ -157,7 +181,8 @@ public final class Database {
     }
 
     /**
-     * Acquires a transaction that can read and write to the database.
+     * Acquires a transaction that can read and write to the database and provides the computed
+     * result.
      *
      * @param handler the handler that is executed within the context of the transaction. The
      *        handler will be called once and its return value is returned from the transaction.
@@ -165,7 +190,7 @@ public final class Database {
      * @return the object that is returned by the given handler
      * @throws DatabaseException if an error occurs in the given handler function
      */
-    public <T> T writeTransaction(
+    public <T> T writeTransactionAndProvide(
             CheckedFunction<? super DSLContext, T, DataAccessException> handler) {
         var holder = new ResultHolder<T>();
 
@@ -190,7 +215,7 @@ public final class Database {
      */
     public void writeTransaction(
             CheckedConsumer<? super DSLContext, ? extends DataAccessException> handler) {
-        writeTransaction(dsl -> {
+        writeTransactionAndProvide(dsl -> {
             handler.accept(dsl);
             // noinspection ReturnOfNull
             return null;
@@ -211,4 +236,3 @@ public final class Database {
         private T result;
     }
 }
-
