@@ -1,63 +1,87 @@
 package org.togetherjava.tjbot.formatter.tokenizer;
 
-import static org.junit.jupiter.api.Assertions.*;
-
-import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
 
-import java.util.ArrayList;
 import java.util.List;
 
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class LexerTest {
-    Lexer lexer;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-    @BeforeAll
-    void init() {
+final class LexerTest {
+    private Lexer lexer;
+
+    @BeforeEach
+    void setUp() {
         lexer = new Lexer();
     }
 
-    @Test
-    @DisplayName("Invalid input does throw an exception")
-    void testInvalidInput() {
-        assertThrows(TokenizationException.class, () -> lexer.tokenize("\r"));
+    private List<TokenType> tokenize(String code) {
+        return lexer.tokenize(code).stream().map(Token::type).toList();
     }
 
     @Test
-    @DisplayName("Valid input does not throw an exception")
-    void testValidInput() {
-        assertDoesNotThrow(() -> lexer.tokenize("\n"));
+    void basic() {
+        String realCode = """
+                int x=5;
+                String y =foo("bar");""";
+
+        List<TokenType> expectedTypes = List.of(TokenType.INT, TokenType.WHITESPACE,
+                TokenType.IDENTIFIER, TokenType.ASSIGN, TokenType.NUMBER, TokenType.SEMICOLON,
+                TokenType.WHITESPACE, TokenType.IDENTIFIER, TokenType.WHITESPACE,
+                TokenType.IDENTIFIER, TokenType.WHITESPACE, TokenType.ASSIGN, TokenType.IDENTIFIER,
+                TokenType.OPEN_PARENTHESIS, TokenType.STRING, TokenType.CLOSE_PARENTHESIS,
+                TokenType.SEMICOLON);
+
+        List<TokenType> actualTypes = tokenize(realCode);
+
+        assertEquals(expectedTypes, actualTypes);
     }
 
     @Test
-    @DisplayName("Empty input does not throw an exception")
-    void testEmptyInput() {
-        assertDoesNotThrow(() -> lexer.tokenize(""));
+    void emptyCode() {
+        String emptyCode = "";
+
+        List<Token> tokens = lexer.tokenize(emptyCode);
+
+        assertTrue(tokens.isEmpty());
     }
 
     @Test
-    @DisplayName("All tokens get tokenized correctly")
-    void testInput() {
-        String testString =
-                "class enum record interface import if for while null extends implements new return package this yield catch try else if else super public private protected static sealed non-sealed final abstract void int long short byte boolean float double char ( ) { } [ ] ; :: : , -> += -= *= /= %= &= |= == != >= > <= < ! = ++ + -- - && || /* a */ / * . ? @annotation 0 \"string\" identifier ` ";
-        List<TokenType> tokens = lexer.tokenize(testString).stream().map(Token::type).toList(),
-                expected = new ArrayList<>();
+    @DisplayName("Keywords can be used as identifier if a letter follows them, such as new vs newText")
+    void keywordVersusIdentifier() {
+        String code = "new newText";
+        List<TokenType> expectedTypes =
+                List.of(TokenType.NEW, TokenType.WHITESPACE, TokenType.IDENTIFIER);
 
-        for (TokenType type : TokenType.values()) {
-            if (type != TokenType.WHITESPACE) {
-                expected.add(type);
-                expected.add(TokenType.WHITESPACE);
-            }
-        }
+        List<TokenType> actualTypes = tokenize(code);
 
-        assertEquals(expected, tokens);
+        assertEquals(expectedTypes, actualTypes);
+    }
+
+    @ParameterizedTest
+    @EnumSource(TokenType.class)
+    @DisplayName("Each token must be recognized by the expected content and not hidden by another token that prefixes it.")
+    void typesAreNotHidden(TokenType expectedTokenType) {
+        String text = expectedTokenType.getContentExample();
+        TokenType actualTokenType = tokenize(text).getFirst();
+
+        assertEquals(expectedTokenType, actualTokenType, "Tested on: " + text);
     }
 
     @Test
-    @DisplayName("Identifier doesn't get recognized as keyword")
-    void testIdentifierKeyword() {
-        assertEquals(TokenType.IDENTIFIER, lexer.tokenize("format").get(0).type());
+    @DisplayName("Nested generics with multi-close > must be identified as GREATER.")
+    void nestedGenericsMultiCloseGreater() {
+        String code = "List<List<Foo>>";
+        List<TokenType> expectedTypes = List.of(TokenType.IDENTIFIER, TokenType.LESS_THAN,
+                TokenType.IDENTIFIER, TokenType.LESS_THAN, TokenType.IDENTIFIER,
+                TokenType.GREATER_THAN, TokenType.GREATER_THAN);
+
+        List<TokenType> actualTypes = tokenize(code);
+
+        assertEquals(expectedTypes, actualTypes);
     }
 }
